Introduce DEPLOY.md - initial draft
Introduce `DEPLOY.md` as the companion to the book `Deploying Reactive Microservices: Strategies for Delivering Resilient Systems`.

The page will describe how to deploy Lagom Chirper example on ConductR sandbox, as well as providing tour of the features provided by ConductR.

This PR is done as part of the first draft to match the deadline of the first draft of the book.