Introduce Maven based build
This PR aims to introduce Maven based build for Chirper. This change is required as part of the O'Reilly book "Deploying Reactive Microservices". The subsequent PR after this change will attempt to introduce native Maven plugin to deploy Lagom based services into ConductR.

**Structure of Changes**

The `pom.xml` files for this build are retrofitted into Chirper from the Lagom `1.3.2` Maven project generator.

The structure of the change is done based on the following Online Auction PR:

https://github.com/lagom/online-auction-java/pull/78

** Removal of SBT build files **

Chirper example includes the `front-end` module which is essentially a Play app. Due to the different structure between Maven based vs SBT based Play app, it is not possible for SBT and Maven build to co-exist. Because of this the SBT related files are now removed.

** Travis Build **

Travis build has now been converted to use Maven with this PR.

The `assembly/conductr-bundle.xml` are introduced to the bundles which are meant to be packaged as an application. This is required since Travis will invoke `maven install` prior executing the maven tasks defined in the `.travis.yml`. Without the presence of `assembly/conductr-bundle.xml`, the `maven install` task will fail causing Travis build to fail.

The generated `.zip` file configured by `assembly/conductr-bundle.xml` only has `lib` directory which contains all the dependent jar files. It is expected that native ConductR Maven plugin to generate the required directories, `bundle.conf`, and executable file.

** Lagom based service changes **

The Lagom based service can be converted with minimal change.

** `front-end` module changes **

The largest change is done on the `front-end` module.

The controller, assets, and views are moved to match the directory structure expected for Maven-based Play project. The `routes` are now described with `Routes.java`.

Compilation of the `.jsx` files requires `node` and `npm` installed through the `frontend-maven-plugin`:

https://github.com/eirslett/frontend-maven-plugin

The `frontend-maven-plugin` provides the means to invoke `npm` to install the `react-tools` which contains the `jsx` executable to compile the `.jsx` files. The resulting `.js` files are then served via `Assets` controller declared in `Routes.java`.

The use of the `frontend-maven-plugin` also means that hot reloading also works with `.jsx` file changes. The compilation of the `.jsx` file is done at the `generate-sources` Maven phase, and the `.jsx` files will be recompiled with every hot reload.

** Note on `react-tools` `npm` module **

The `react-tools` is deprecated in lieu of the `babeljs` compiler: http://babeljs.io/

The `babeljs` compiler is able to compile many variants of Javascripts (i.e. ES5, ES6, JSX) into browser compatible Javascript.

`babeljs` itself pulls quite a large number of dependencies, causing the `npm install` and `.jsx` compilation step to take a significantly longer time compared to the use of `react-tools`.

Because of this reason the `react-tools` is chosen as it doen't impede the time it takes for the hot reload to complete.